requirements.txt
pandas
numpy
scikit-learn
hdbscan
pdfplumber
openpyxl
requests
pyarrow
PyPDF2
transformers
torch
sentence-transformers
chromadb
langchain
langchain-community
langchain-openai
langchain-google-genai
python-dotenv
pydantic-settings
langgraph
fastapi
uvicorn
python-multipart
sqlalchemy